[#151] Refactor the existing modules
