+++ src/templates/aws/addons/core/common.test.ts
@@ -1,5 +1,5 @@
-import { AwsOptions } from '..';
-import { remove } from '../../../helpers/file';
+import { AwsOptions } from '../../';
+import { remove } from '../../../../helpers/file';
 import applyCommon from './common';
 
 describe('Common add-on', () => {
