+++ src/templates/aws/addons/index.ts
@@ -1,15 +1,15 @@
 import applyAlb from './alb';
 import applyBastion from './bastion';
 import applyCloudwatch from './cloudwatch';
-import applyCommon from './common';
+import applyCommon from './core/common';
+import applyRegion from './core/region';
+import applySecurityGroup from './core/securityGroup';
+import applyVpc from './core/vpc';
 import applyEcr from './ecr';
 import applyEcs from './ecs';
 import applyRds from './rds';
-import applyRegion from './region';
 import applyS3 from './s3';
-import applySecurityGroup from './securityGroup';
 import applySsm from './ssm';
-import applyVpc from './vpc';
 
 export {
   applyAlb,
