+++ src/templates/aws/advanced.test.ts
@@ -3,16 +3,12 @@ import { remove } from '../../helpers/file';
 import {
   applyAlb,
   applyBastion,
-  applyCommon,
   applyEcr,
   applyEcs,
   applyCloudwatch,
   applyRds,
-  applyRegion,
   applyS3,
-  applySecurityGroup,
   applySsm,
-  applyVpc,
 } from './addons';
 import { applyAdvancedTemplate } from './advanced';
 
@@ -35,22 +31,6 @@ describe('AWS advanced template', () => {
     afterAll(() => {
       jest.clearAllMocks();
       remove('/', projectDir);
-    });
-
-    it('applies common add-on', () => {
-      expect(applyCommon).toHaveBeenCalledWith(options);
-    });
-
-    it('applies region add-on', () => {
-      expect(applyRegion).toHaveBeenCalledWith(options);
-    });
-
-    it('applies VPC add-on', () => {
-      expect(applyVpc).toHaveBeenCalledWith(options);
-    });
-
-    it('applies security group add-on', () => {
-      expect(applySecurityGroup).toHaveBeenCalledWith(options);
     });
 
     it('applies ECR add-on', () => {
