+++ src/templates/aws/advanced.ts
@@ -2,23 +2,15 @@ import { AwsOptions } from '.';
 import {
   applyAlb,
   applyBastion,
-  applyCommon,
   applyEcr,
   applyEcs,
   applyCloudwatch,
   applyRds,
-  applyRegion,
   applyS3,
-  applySecurityGroup,
   applySsm,
-  applyVpc,
 } from './addons';
 
 const applyAdvancedTemplate = (options: AwsOptions) => {
-  applyCommon(options);
-  applyRegion(options);
-  applyVpc(options);
-  applySecurityGroup(options);
   applyEcr(options);
   applyCloudwatch(options);
   applyS3(options);
